Use "dtls_cert_file" config option in respoke.conf
Previously, the Dockerfile and builder script were using
"dtls_private_key" configuration option to set the cert location needed
for dtls. However, when using this option, the SDP generated when
sending an answer to an incoming call had an empty fingerprint hash,
which is invalid for WebRTC.

This patch changes this config option to "dtls_cert_file", which has
been verified to work correctly for web->chan_respoke calling.